Add smart enemy, change coordinates to float

+++ src/com/kmne68/main/FastEnemy.java
@@ -21,7 +21,7 @@ public class FastEnemy extends GameObject {
 	
 	public Rectangle getBounds() {
 		
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle( (int) x, (int) y, 16, 16);
 		
 	}
 	
@@ -37,7 +37,7 @@ public class FastEnemy extends GameObject {
 		if(x <= 0 || x >= Game.WIDTH - 32)
 			velocityX *= -1;
 		
-		handler.addObject(new Trail( x, y, ID.Trail, Color.cyan, 16, 16, 0.02f, handler));
+		handler.addObject(new Trail( (int) x, (int) y, ID.Trail, Color.cyan, 16, 16, 0.02f, handler));
 		
 	}
 	
@@ -46,7 +46,7 @@ public class FastEnemy extends GameObject {
 	public void render(Graphics g) {
 
 		g.setColor(Color.cyan);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect( (int) x, (int) y, 16, 16);
 		
 	}
